video_gaze_estimation/requirements.txt
click==8.1.5
Flask==1.0.2
gevent==1.5.0
greenlet==2.0.2
gunicorn==20.0.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
numpy==1.20.1
onnxruntime-gpu==1.4.0
opencv-python==4.5.1.48
protobuf==4.23.4
py-eureka-client==0.5.7
PyYAML==5.3.1
requests==2.12.4
scipy==1.6.1
typing_extensions==4.7.1
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
Pillow